Refactoring: Create CourseCompletionReportRequestService
